+++ financial_loss_functions/src/data_processing/dataset.py
@@ -235,39 +235,6 @@ class WindowDataset(Dataset):
         # Return one sample
         return self.X[idx], self.y[idx]
 
-
-def calc_in_out_idx(
-        split_data: pd.DataFrame, in_size: int, out_size: int, stride: int
-    ) -> tuple[list[tuple], list[tuple]]:
-
-    #### Implment start index shifting of training data here, if needed ####
-    #### Current design intended to use entire train data.
-
-    # Check for missing data
-    if split_data.isna().any().any():
-        raise ValueError('Split has missing data. Fix before training.')
-    
-    starts = list(
-        range(0, len(split_data) - (in_size + out_size) + 1, stride)
-    )
-
-    in_sample_indexes = []
-    out_sample_indexes = []
-    for strt in starts:
-        in_end = strt + in_size
-        in_sample_indexes.append((strt, in_end))
-        
-        # FIX: out_start must be exactly in_end
-        out_start = in_end 
-        out_end = out_start + out_size
-        out_sample_indexes.append((out_start, out_end))
-
-    if len(in_sample_indexes) != len(out_sample_indexes):
-        raise RuntimeError('Window count mismatch.')
-
-    return in_sample_indexes, out_sample_indexes
-
-
 class WFAdjustment:
     def __init__(self, out_size: int):
         self.out_size = out_size
